backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.0
aiofiles==23.2.1
jinja2==3.1.2
email-validator==2.1.0
google-generativeai==0.3.2
aiohttp==3.9.1

# Discord Integration
discord.py==2.3.2

# Analytics and Monitoring
redis==5.0.1
aioredis==2.0.1
prometheus-client==0.19.0
psutil==5.9.6

# WebSocket support
websockets==12.0

# Database
databases[postgresql]==0.8.0
asyncpg==0.29.0

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Security
cryptography==41.0.8
pycryptodome==3.19.0

# Environment and configuration
pydantic-settings==2.1.0
alembic==1.12.1
bcrypt==4.1.2
